requirements.txt
requests
polars
python-dotenv
tqdm
beautifulsoup4
lxml
readability-lxml
